Supprimer Salarie

+++ Client.cs
@@ -10,7 +10,7 @@ namespace Projet_TransConnect
     {
          
 
-        public Client(int id, string prenom, string nom, DateTime naissance, string adresse, string mail, int telephone) : base(id, prenom, nom, naissance, adresse, mail, telephone)
+        public Client(int id, string prenom, string nom, DateTime naissance, string adresse, string mail, long telephone) : base(id, prenom, nom, naissance, adresse, mail, telephone)
         {
             
         }
